For helper

+++ app/controllers/follows_controller.rb
@@ -1,18 +1,23 @@
 class FollowsController < ApplicationController
   before_action :authenticate_user!
+  before_action :set_users
 
-  def create
-    follower_user = current_user
-    following_user = User.find(params[:following_id])
-    Follow.create(follower: follower_user, following: following_user)
-    redirect_to user_followings_path(current_user)
-  end
+    def create
+      Follow.create(follower: @follower_user, following: @following_user)
+      redirect_back(fallback_location: user_followings_path(@follower_user))
+    end
 
   def destroy
-    follower_user = current_user
-    following_user = User.find(params[:following_id])
-    follow = Follow.find_by(follower: follower_user, following: following_user)
-    follow.destroy
-    redirect_to user_followers_path(current_user)
+    if follow = Follow.find_by(follower: @follower_user, following: @following_user)
+      follow.destroy
+    end
+    redirect_back(fallback_location: user_followers_path(@follower_user))
+  end
+
+  private
+
+  def set_users
+    @follower_user = current_user
+    @following_user = User.find(params[:following_id])
   end
 end
